// src/pages/thunder-sdk/src/types/erc721/NFTContractAbi.d.ts
export type IdentityInput = Enum<{ Address: AddressInput, ContractId: ContractIdInput }>;
export type IdentityOutput = Enum<{ Address: AddressOutput, ContractId: ContractIdOutput }>;
export type MetadataInput = Enum<{ B256: string, Bytes: BytesInput, Int: BigNumberish, String: StringInput }>;
export type MetadataOutput = Enum<{ B256: string, Bytes: BytesOutput, Int: BN, String: StringOutput }>;
// ...
export type AddressInput = { value: string };
export type AddressOutput = AddressInput;
export type AssetIdInput = { value: string };
export type AssetIdOutput = AssetIdInput;
export type BytesInput = { buf: RawBytesInput, len: BigNumberish };
export type BytesOutput = { buf: RawBytesOutput, len: BN };
export type ContractIdInput = { value: string };
export type ContractIdOutput = ContractIdInput;
export type RawBytesInput = { ptr: BigNumberish, cap: BigNumberish };
export type RawBytesOutput = { ptr: BN, cap: BN };
export type StringInput = { bytes: BytesInput };
export type StringOutput = { bytes: BytesOutput };

interface NFTContractAbiInterface extends Interface {
  functions: {
    decimals: FunctionFragment;
    name: FunctionFragment;
    symbol: FunctionFragment;
    total_assets: FunctionFragment;
    total_supply: FunctionFragment;
    burn: FunctionFragment;
    mint: FunctionFragment;
    metadata: FunctionFragment;
    set_decimals: FunctionFragment;
    set_name: FunctionFragment;
    set_symbol: FunctionFragment;
    set_metadata: FunctionFragment;
  };

  encodeFunctionData(functionFragment: 'decimals', values: [AssetIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: 'name', values: [AssetIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: 'symbol', values: [AssetIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: 'total_assets', values: []): Uint8Array;
  encodeFunctionData(functionFragment: 'total_supply', values: [AssetIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: 'burn', values: [string, BigNumberish]): Uint8Array;
  encodeFunctionData(functionFragment: 'mint', values: [IdentityInput, string, BigNumberish]): Uint8Array;
  encodeFunctionData(functionFragment: 'metadata', values: [AssetIdInput, StringInput]): Uint8Array;
  encodeFunctionData(functionFragment: 'set_decimals', values: [AssetIdInput, BigNumberish]): Uint8Array;
  encodeFunctionData(functionFragment: 'set_name', values: [AssetIdInput, StringInput]): Uint8Array;
  encodeFunctionData(functionFragment: 'set_symbol', values: [AssetIdInput, StringInput]): Uint8Array;
  encodeFunctionData(functionFragment: 'set_metadata', values: [AssetIdInput, StringInput, MetadataInput]): Uint8Array;

  decodeFunctionData(functionFragment: 'decimals', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'name', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'symbol', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'total_assets', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'total_supply', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'burn', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'mint', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'metadata', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'set_decimals', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'set_name', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'set_symbol', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'set_metadata', data: BytesLike): DecodedValue;
}

export class NFTContractAbi extends Contract {
  interface: NFTContractAbiInterface;
  functions: {
    decimals: InvokeFunction<[asset: AssetIdInput], Option<number>>;
    name: InvokeFunction<[asset: AssetIdInput], Option<StringOutput>>;
    symbol: InvokeFunction<[asset: AssetIdInput], Option<StringOutput>>;
    total_assets: InvokeFunction<[], BN>;
    total_supply: InvokeFunction<[asset: AssetIdInput], Option<BN>>;
    burn: InvokeFunction<[sub_id: string, amount: BigNumberish], void>;
    mint: InvokeFunction<[recipient: IdentityInput, sub_id: string, amount: BigNumberish], void>;
    metadata: InvokeFunction<[asset: AssetIdInput, key: StringInput], Option<MetadataOutput>>;
    set_decimals: InvokeFunction<[asset: AssetIdInput, decimals: BigNumberish], void>;
    set_name: InvokeFunction<[asset: AssetIdInput, name: StringInput], void>;
    set_symbol: InvokeFunction<[asset: AssetIdInput, symbol: StringInput], void>;
    set_metadata: InvokeFunction<[asset: AssetIdInput, key: StringInput, metadata: MetadataInput], void>;
  };
}